// src/app/models/data-consolidate-asset.ts
export interface DataConsolidateAsset {
	productId: number;
	productName: string;
	equity: number;
	portfolioPercentage: number;
	productTypeId: number;
	productTypeName: string;
	financialInstitutionId: number;
	financialInstitutionName: string;
	strategyOfDiversificationId: number;
	strategyOfDiversificationDescription: string;
}